Remove server.js demo
